LayoutTests/http/tests/websocket/tests/hybi/upgrade-simple-ws-expected.txt

Simple upgrade insecure Web Socket test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Created a socket to 'wss://127.0.0.1:9323/websocket/tests/hybi/simple'; readyState 0.
The 'ws' protocol should be upgraded to 'wss'. If this happens, we will connect properly and the test will succeed.
Connected; readyState 1
Resolved URL: wss://127.0.0.1:9323/websocket/tests/hybi/simple
Received: 'Hello from Simple WSH.'; readyState 1
Closed; readyState 3.
